// src/app/itaap-ms-manager/api/enums.ts
export enum Status {
  N = 'new',
  A = 'accepted',
  R = 'rejected',
  P = 'inprogress',
  C = 'completed',
  O = 'overdue',
  B = 'blocked'
}

export enum Role {
  A = 'Admin',
  T = 'Tech Lead',
  D = 'Developer',
  M = 'Scrum Master',
  R = 'Read',
  C = 'Client'
}

export enum Priority {
  C = 'Critical',
  H = 'High',
  M = 'Medium',
  L = 'Low',
}